Finish refactoring transformers
